pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nk-russia-tnr-tracker"
version = "0.1.0"
description = "NK-Russia Human Rights Chain of Command Tracker"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "TNR Tracker Team"}
]
keywords = ["human-rights", "north-korea", "russia", "sanctions", "tnr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.25",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]
pdf = [
    "marker-pdf>=0.2.0",
    "pymupdf>=1.23.0",
]

[project.scripts]
tnr-tracker = "packages.api.main:app"
tnr-etl = "packages.etl.pipeline:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["packages*"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=packages --cov-report=term-missing"
